+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes, including bugfixes, can be found here.
 
+## Release v1.1.1
+
+- Fixed image in readme.
+- Fixed bug where system libraries were getting indexed.
+- Added timeout to norminette.
+
 ## Release v1.1.0
 
 - Added error highlighting for makefiles.